moving line profiler instrument tests in a separate file
